+++ indexers/gemini-3h/consensus/src/mappings/types.ts
@@ -1,0 +1,19 @@
+type ExtrinsicPrimitive = {
+  callIndex: string;
+  args: any;
+};
+
+export type ExtrinsicHuman = ExtrinsicPrimitive & {
+  method: string;
+  section: string;
+};
+
+export type EventPrimitive = {
+  index: string;
+  data: any;
+};
+
+export type EventHuman = EventPrimitive & {
+  method: string;
+  section: string;
+};
